+++ flutter/lib/src/interface/connection_handle.dart
@@ -1,6 +1,5 @@
 import '/src/util/rust_handles_storage.dart';
-import 'remote_media_track.dart';
-import 'track.dart';
+import 'media_track.dart';
 
 /// External handler to a `Connection` with a remote `Member`.
 abstract class ConnectionHandle implements PlatformHandle {
