--- flutter/lib/src/interface/local_media_track.dart
@@ -1,8 +1,0 @@
-import 'track.dart';
-
-/// Strongly referenced media track received from a
-/// [`getUserMedia()`][1]/[`getDisplayMedia()`][2] request.
-///
-/// [1]: https://w3.org/TR/mediacapture-streams#dom-mediadevices-getusermedia
-/// [2]: https://w3.org/TR/screen-capture#dom-mediadevices-getdisplaymedia
-abstract class LocalMediaTrack implements MediaTrack {}
