+++ flutter/lib/src/web/connection_handle.dart
@@ -3,8 +3,7 @@ import 'dart:js';
 import 'package:js/js.dart';
 
 import '../interface/connection_handle.dart';
-import '../interface/remote_media_track.dart';
-import '../interface/track.dart';
+import '../interface/media_track.dart';
 import '../util/move_semantic.dart';
 import 'exceptions.dart';
 import 'jason_wasm.dart' as wasm;
