+++ flutter/lib/src/web/remote_media_track.dart
@@ -2,8 +2,7 @@ import 'package:js/js.dart';
 import 'package:medea_flutter_webrtc/medea_flutter_webrtc.dart' as webrtc;
 import 'package:medea_flutter_webrtc/src/platform/web/media_stream_track.dart';
 
-import '../interface/remote_media_track.dart';
-import '../interface/track.dart';
+import '../interface/media_track.dart';
 import '../util/move_semantic.dart';
 import 'exceptions.dart';
 import 'jason_wasm.dart' as wasm;
